requirements.txt
os
sys
shutil
pygithub
threading
pyinstaller
python-dotenv